viced: check if pt connections are out of date
The connections from the fileserver to the pt server are created
only once when the fileserver is initialized. As a result, if the
ip address of the pt server is modified, every single fileserver
thread will have to timeout in order to create a new connection.
To avoid this problem, the fileserver have to be restarted every
time the ip address of the pt server is modified. Unfortunately,
this workaround is not always desirable.

This patch assumes that these connections should be recreated every
time the configuration file is modified. As a result, the connections
from the fileserver to the pt server will be always up to date. To make
it possible, every single fileserver thread needs to keep its own copy
of the configuration file that was used to create these connections.