DARWIN: Fix removing dirs in Finder with 10.11
Starting with OS X 10.11, if you try to remove a directory in Finder,
a process called DesktopServicesHelper will try to access the dir.
That process runs as root, and so if root does not have access to the
relevant path, we'll fail with EACCES, which will prevent Finder from
removing the dir.

To this issue, ignore some afs_vop_access EACCES errors for root. It
is difficult or impossible to tell if the accessing process is
DesktopServicesHelper, and what user it's requesting access for, so
just ignoring the error for all root processes is the best we can do.
Only do this for 10.11 and on, only for PRSFS_LOOKUP requests, and
only for directories, to try to restrict this behavior to the fewest
situations possible while still avoiding the issue.

This should not be a huge security concern, since it's just for root
processes. However, this change does allow root processes to stat(2)
any directory, regardless of permissions, if that directory is cached.
This arguably leaks data to root, but similar data can already be
retrieved from cmdebug and by inspecting the on-disk cache files
directly.

A similar problem with DesktopServicesHelper is noted in the osxfuse
kext.git repository, commit 80096780a818bdfc8d18ba3f3d36a17fb83f43eb.

Change-Id: Ib4fb154fc70388cabd31f87436e1cb9f6ec5ebb7