util: resolve and cache the hostnames already resolved
Generally speaking, every time the hostname of a specific IP
address is needed, the hostutil_GetNameByINet() function is
called. Often, this function is called several times to resolve
the same IP address. This patch provides a caching mechanism to
store the hostnames of the IP addresses already resolved. Moreover,
the function introduced by this patch is compatible with IPv6
addresses.

This caching can significantly speed up vos listvldb, since the
vldb typically has a large number of vldb entries but a very
small number of distinct file server addresses.

Notes:
  - Some of the ideas of this patch are based on gerrit 11390.
  - Assuming that LWP will be completely removed in the future,
    these new files can be moved to opr.