volser: vos listvldb caching dns results
When vos listlvdb is called, every single volume, along with
the hostname of the server responsible for it, is printed out.
Often, the same server stores a huge number of volumes. As a
result, the same IP address needs to be resolved multiple times.

This patch improves the performance of vos listvldb, caching
the hostnames of the IP addresses already resolved. To do so,
the function SubEnumerateEntry calls util_gethostname instead
of hostutil_GetNameByINet.