macos: use pkgbuild to build the package on 10.10/10.11
PackageMaker is no longer part of OS X. As a result, it
is not possible to build the package on OS X 10.10 and
OS X 10.11 using the existing code.

To solve this problem, a new script, along with a couple
of new files, are provided.

- pkgbuild.sh

This script uses the command line tools pkgbuild and
productbuild to build the package on OS X 10.10 and
OS X 10.11. By default, the package built by this
script will not be signed. Optionally, the package
might be signed.

- Distribution.xml

This file is nothing more than an XML file used by
productbuild. It is mainly used to configure how the
installer is gonna look and behave.

- conclusion.txt

The message provided by this file will be launched
at the very end of the installation process.

- Uninstall.14+

This script can be used by OS X 10.10/10.11 users
to uninstall OpenAFS.

Notes:

    - El Capitan and further prevent us from touching
      /usr/bin directly. As a result, /opt is used.
    - If the package is not signed, the user will have
      to disable Mac security protections on his OS X.
      Otherwise, the client will not work.
    - Now we have two different scripts to build the
      package on OS X. For OS X 10.10 and newer versions,
      pkgbuild.sh will be used. For older versions,
      the usual buildpkg.sh will be used.