afs: afs_pag_wait() makes process unkillable
The afs_pag_wait() function is called every time we need to
create a new pag. If the caller is not a super user, this
function is gonna sleep until the point where the difference
between the current time and pag_epoch gets greater than the
pagCounter. Unfortunaly, the function in question ignores
the code provided by afs_osi_Wait(). As a result, it is not
possible to kill the process that requested the new pag while
afs_pag_wait() is still slepping. In fact, if the user tries
to kill this process it will spin the cpu.

To fix this problem, do not ignore the code provided by the
function afs_osi_Wait().