venus: fix memory leak
The server ranks might be requested through the function
pioctl. If the size of the output buffer is not enough,
the function in question will return the code E2BIG.

The server ranks are requested in the function GetPrefCmd.
If the code E2BIG is received, we allocate more memory and
try again. However, if the size of the output buffer reaches
16384 bytes and this space is still not enough, we return
without release the memory that was previously allocated.

To fix this problem, do not return without release the
output buffer.