sys: retry lsetpag if it was interrupted
The variable errno might be set by some system calls to
indicate the reason why the system call in question did
not work as expected. If the system call lsetpag is
interrupted by a signal, the value of the variable errno
will be equal to EINTR. This value means that the system
call did not succeed because it was interrupted.

If the system call lsetpag did not succeed because it was
interrupted, try again.