sys: retry lsetpag if errno is equal to EINTR
The variable errno might be set by some system calls to
indicate the reason why the system call in question did
not work as expected. If the setpag system call is
interrupted by a signal, the value of the variable errno
will be equal to EINTR. This value means that setpag
did not succeed because it was interrupted.

If lsetpag did not succeed and errno is equal to EINTR,
try again.

Change-Id: Ibf306d62fc8d2fa9ccb0692f9031c5aa659b2bfe