ubik: add libubik capability to use kernel preferences
In a scenario with multiple db servers, the cache manager
needs to know the order in which it should contact the vl
servers. To make it possible, the cache manager stores a
preference rank in kernel memory.

On the other hand, the clients of some authentication
services do not have the same resource. No rank is taken
into consideration when the client needs to contact
processes like kaserver and pt server. Instead, a server
is chosen randomly.

The goal of this patch is to add libubik capability to
use kernel preferences. As a result, clients of services
like the ones mentioned above will have the option to use
the same kernel preferences used by the cache manager.