aklog: use kernel db server preferences
In a scenario with multiple db servers, aklog contacts a
pt server chosen at random from the cell’s entry in the
local CellServDB file. This approach is not always
desirable.

The goal of this patch is to make aklog use the same
kernel db server preferences used by the cache manager.
Optionally, this feature might be turned off.