aklog: make klog.krb5 use kernel db server preferences
In a scenario with multiple db servers, klog.krb5 contacts a
pt server chosen at random from the cell’s entry in the
local CellServDB file. This approach is not always
desirable.

The goal of this patch is to make klog.krb5 use the same
kernel db server preferences used by the cache manager.
Optionally, this feature might be turned off.