kauth: make klog use kernel db server preferences by default
In a scenario with multiple kaservers, klog contacts an
authentication server chosen at random from the cell's
entry in the local CellServDB file. This approach is not
always desirable.

The goal of this patch is to make klog use by default the
same kernel db server preferences used by the cache manager.
Optionally, this feature might be turned off.

Note: This work is based on a project originally developed
by Mark Vitale <mvitale@sinenomine.net>.

Change-Id: Ia11a7f3c27b1379c7e291f4233c51c20a99f7c10