kauth: make klog use kernel db server preferences by default
In a scenario with multiple kaservers, klog contacts an
authentication server chosen at random from the cell's
entry in the local CellServDB file. This approach is not
always desirable.

Make klog use the cache manager’s kernel db server
preferences by default. If the kernel prefs are not
available, or if the new -nokernelprefs option is
specified, use the former method of random server
selection.

Note: This work is based on a project originally developed
by Mark Vitale <mvitale@sinenomine.net>.

Change-Id: Ia11a7f3c27b1379c7e291f4233c51c20a99f7c10