pam: use kernel db server preferences
In a scenario with multiple kaservers, pam contacts an
authentication server chosen at random from the cell's
entry in the local CellServDB file. This approach is not
always desirable.

The goal of this patch is to make pam use the same kernel
db server preferences used by the cache manager.

Note: This work is based on a project originally developed
by Mark Vitale <mvitale@sinenomine.net>.

Change-Id: I472dbab272038fbefbcdcf8b8afd17185bc638b7