aklog: use kernel db server preferences
In a scenario with multiple db servers, aklog contacts a
pt server chosen at random from the cell’s entry in the
local CellServDB file. This approach is not always
desirable.

Make aklog use the cache manager’s kernel db server
preferences by default. If the kernel prefs are not
available, or if the new -nokernelprefs option is
specified, use the former method of random server
selection.

Note: This work is based on a project originally developed
by Mark Vitale <mvitale@sinenomine.net>.

Change-Id: Ie322d707fed6eae8bc32b2da2e5f23e39f1bf41f