vol: add capability to add partitions at runtime
Currently, it is not possible to add new partitions to the servers at
runtime. If a new vice partition is added to the system, the servers
will have to be restarted.

To make it possible, this patch introduces a couple of functions that
might be used at runtime to add and remove empty partitions.