vol: check if partition can be attached at runtime
The VAttachPartitions() function finds the vice partitions on the
server. In order to check if every single partition found is valid,
the VCheckPartition() function is called. This function performs some
basic checks and, if the partition in question is valid, it will be
added to the DiskPartition list and to the DiskPartition table.

Since these functions are only used during the initialization of the
server, they do not check if the partition in question is already
attached. However, this commit introduces a new feature that allows the
attachment of new partitions at runtime. To make it possible, the
VAttachPartitions() function might be called several times.

In order to avoid multiple attachments of the same partition, make
VCheckPartition() ignores the partitions already attached. Also, make
sure that the new partition does not have any volume.