ubik: drop versionLock when DISK_Begin is called
When LWP is being used, the versionLock is not dropped when the sync
site is waiting for the response of DISK_Begin. As a result, any
subsequent transaction will be blocked by this lock during this time.

In order to allow read transactions while a write transaction is waiting
for the response of DISK_Begin, drop the versionLock when this RPC is
called and reacquire when it returns.