ubik: acquire remote rwlock before local rwlock
Currently the sync site only calls the DISK_Lock RPC when the local
rwlock is already acquired. As a result, any subsequent read transaction
will be blocked by this lock while the sync site is waiting for the
response of DISK_Lock.

In order to allow read transactions while a write transaction is waiting
for the response of DISK_Lock, reverse the order in which the rwlock is
acquired.