setup: include helper script and remove i386 arch
This commit includes a script called osx_builder.sh that can be used to
build the code on Sierra. Also, the i386 architecture had to be removed
from the xcode project.

The osx_builder.sh script can be used to build the code. After that, run
the following command as root to generate the installer:

$ sudo ARCHFLAGS="-arch x86_64" make packages