afsd: add an option to use csdb as a whitelist
If a nonexistent file under the /afs directory is referenced when the
-dynroot and -afsdb options are enabled, afs will see this file as a
possible cell. As a result, afsdb will try to resolve it. This behavior
is not always desirable.

To fix this problem, add an option to use csdb as a whitelist. As a
result, the cells that cannot be found in the whitelist will not be
resolved by afsdb.