opr: backport basic libraries from master
The master branch provides a set of basic libraries already used by
some patches developed for this branch. These libraries are not
available on 1.6 yet. As a result, we cannot backport the patches in
question.

To fix this problem, backport opr/jhash, opr/queue, opr/lock, and
opr/dict from master.