ubik: don't start writes if remotes don't have sync-site
There is a gap between the time when a new sync-site is elected and the
time when the remotes are aware of that. As expected, any write
transaction between this gap will fail. This will force a new election
which might be time consuming.

The problem in question can be exacerbated if we have a heavy load of
write transactions at the moment when quorum is lost. In this scenario,
the writes in the gap mentioned above will force several re-elections.
As a result, the dbserver will not be able to write any information for
a long time.

To fix this problem, only start write transactions if at least quorum
are aware we have a sync-site.