ubik: update epoch when sync-site is elected
The ubik_epochTime represents the time at which the coordinator first
received its coordinator mandate. However, this global is not updated
at the moment when a new sync-site is elected. Instead, ubik_epochTime
is only updated at the very end of the first write transaction. As a
result, at the begin of the first write transaction the ubik_epochTime
will be equal to the epoch of the last sync-site. This epoch will be
sent to the remotes through SDISK_Begin(). When ubik_epochTime is
finally updated at the end of the first write transaction on the
sync-site, the epoch from the sync-site will not be equal to the one
stored by the remotes. If the beacon thread sends a beacon packet to
the remotes before the end of the transaction in question, the remotes
will realize that its local epoch is not equal to the one held by the
sync-site and the transaction will be aborted. This failure might force
a re-election process.

To fix this problem, update ubik_epochTime as soon as the sync-site is
elected.