ubik: allow remote dbase relabel if up to date
When a site is elected the sync-site, its database is not immediately
relabeled. The database in question will be relabeled at the end of the
first write transaction (in udisk_commit). To do so, the dbase->version
is updated on the sync-site first (1) and then the versions of the
remote sites are updated through SDISK_SetVersion() (2).

In order to make sure that the remote site holds the same database as
the sync-site, the SDISK_SetVersion() function checks if the current
version held by the remote site (ubik_dbVersion) is equal to the
original version stored by the sync-site (oldversionp). If
ubik_dbVersion is not equal to oldversionp, SDISK_SetVersion() will not
work.

However, the ubik_dbVersion can be updated by the vote thread at any
time. As a result, if the beacon thread runs between (1) and (2) on the
sync-site, the vote thread on a remote site can set ubik_dbVersion to
the newest version while its ubik_dbase->version is still equal to the
old one. As a result, ubik_dbVersion will not be equal to oldversionp
and SDISK_SetVersion() will not work. This failure may cause a loss of
quorum until another election is completed.

To fix this problem, let SDISK_SetVersion() relabels the database when
ubik_dbase->version is equal to oldversionp. In order to restrict this
change to the scenario described above, also check if ubik_dbVersion is
equal to newversionp.