ubik: don't start writes before the second voting period
By definition, quorum is established when the sync-site is elected. This
means that, after executing the recovery operation, the new coordinator
should be able to execute write transactions on the database. However,
the introduction of 39729bf290e9ebfa360b79e5999dac01910e341f changed
this original behavior.

In order to fix a different issue, the commit mentioned above introduced
a new sanity check at the begin of SDISK_Begin() (urecovery_AllBetter).
As a result, SDISK_Begin() will return UNOQUORUM if lastYesState is not
set. Since lastYesState is only set on the remotes when the first beacon
packet is received from the new sync-site, any write transaction started
after the recovery operation and before the second voting period will
force the remotes to return UNOQUORUM.

This problem was exposed by fac0b742960899123dca6016f6ffc6ccc944f217.
Without this commit, failures on SDISK_Begin() could be ignored.

The problem in question can be exacerbated if we have a constant load of
write transactions in the cell. In this scenario, the writes before the
second voting period may force several re-elections.

In order to avoid unnecessary re-elections, do not allow write
transactions before the second voting period. In other words, only allow
write transactions when the sync-site is elected and the remote sites
are aware of that.

Change-Id: I9e1a76b4022e6d734af1165d94c12e90af04974d