libafs: add ctf debugging information to solaris
CTF is a reduced form of debug information similar to DWARF and stab. It
describes types and function prototypes. Basically, the ctf tools take
the debugging information generated by the '-g' compiler flag, and strip
out everything but the type information. The principal objective of the
format is to shrink the data size as much as possible so that it could
be included in a production environment. MDB, DTrace, and other tools
use CTF debug information to read and display structures correctly.