ubik: don't set database epoch to 0 if not needed
If our attempt to receive a fresh database from a peer fails, we will
overwrite the version.epoch field of our current local copy of the
database with an invalid value, "0". The idea behind this approach is
to make sure that this database will not be seen as a legit copy if the
transfer is not completed properly. Although it is questionable if this
approach is still necessary (since the current version writes the data
into a temporary file), it is undisputed that the database version does
not have to be invalidated if the transfer fails in a early stage where
no data has been written and we could safely continue to reuse the local
copy for read-only queries. Early failures may happen if:

1. the peer sending the database to us is not the peer we believe to be
the sync site;

2. the call used by the sender is not authenticated;

In both cases, the database epoch is invalidated. As a result of that,
we may have the following consequences:

1. reads may not be allowed

2. data can be lost

If the site in question happens to be the one with the most recent
database, during a new quorum establishment, the database of this site
will not be taken into consideration.

To fix the problem in question, do not set the database epoch to 0
if the local database was not modified.

Acknowledgements:

Hartmut Reuter <hartmut.reuter@gmx.de>
    - found the problem;
    - suggested a possible solution;

Benjamin Kaduk <kaduk@mit.edu>
    - submitted the first version;

Andrew Deason <adeason@dson.org>
    - suggested changes;