ubik: Avoid urecovery_AllBetter check in ubik_Seek
Currently, ubik_Seek checks urecovery_AllBetter to see if we still have
quorum for the given ubik db. This is completely unnecessary, since this
function doesn't interact with the database at all; we only set a couple
of fields in the given transaction structure.

To slightly improve performance, avoid checking urecovery_AllBetter
here. We can also remove the database locks as well, since we are no
longer interacting with the ubik database in any way.

Change-Id: I90b49005a29fea0cf51512caa59157b7da353a9e