autoconf: do not use the missing scrip
The missing script is included by automake. Since OpenAFS does not use
automake, this script is not being copied to the build-tools directory.
As a result, am_missing_run is not initialized.

In order to preserve the expected behavior, this commit replaces
am_missing_run by a function that prints a friendly message when the
required binary is not found.