rx: fix atomics on darwin
As described by commit b2a21, the API used for atomic operations on
kernel space is not the same as the one used on user space. To fix this
problem, b2a21 introduced macros to correct the name of these functions
on kernel space. Unfortunately, the return value of the functions used
on kernel space is not the same as the ones used on user space.
Generally speaking, the kernel space atomic functions return the
original value of the variable received as an argument before the
operation in question. On the other hand, the user space atomic
functions return the new value, after the operation has been performed.
To fix this problem, this commit provides a new set of inline functions
(only used on kernel space) with the expected return values.

Also, in order to get the inline implementations of the OSAtomic
interfaces in terms of the <stdatomic.h> primitives, commit 74f83
defines OSATOMIC_USE_INLINED on OS X 10.12. However, the definition of
this macro only affects the user space legacy interfaces for atomic
operations. The kernel space interfaces for atomics are not deprecated
and OSATOMIC_USE_INLINED does not affect these functions. To fix this
problem, only define OSATOMIC_USE_INLINED on user space (OS X 10.12+).